Refactor utils.js
